project/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
